added shared router tests for feature parity between IttyRouter and Rotuer
